Extract fakes and test helpers
