+++ tests/tests.py
@@ -1,134 +1,11 @@
-import os
-from pathlib import Path
-import subprocess
-from collections import namedtuple
-
-from fixtures import TestWithFixtures, TempDir
-
-
-ProcessResult = namedtuple('ProcessResult', ['returncode', 'stdout', 'stderr'])
-
-SNAP_LIVEPATCH_INSTALLED = """#!/bin/sh
-if [ "$1" = "list" ]; then
-    cat <<EOF
-Name                 Version  Rev  Developer  Notes
-canonical-livepatch  7        22   canonical  -
-EOF
-elif [ "$1" = "install" ]; then
-    cat <<EOF
-snap "canonical-livepatch" is already installed, see "snap refresh --help"
-EOF
-elif [ "$1" = "remove" ]; then
-    echo "canonical-livepatch removed"
-fi
-exit 0
-"""
-
-SNAP_LIVEPATCH_NOT_INSTALLED = """#!/bin/sh
-if [ "$1" = "list" ]; then
-    cat <<EOF
-error: no matching snaps installed
-EOF
-    exit 1
-elif [ "$1" = "install" ]; then
-    cat <<EOF
-canonical-livepatch 7 from 'canonical' installed
-EOF
-    exit 0
-fi
-"""
-
-LIVEPATCH_ENABLED = """#!/bin/sh
-if [ "$1" = "status" ]; then
-    cat <<EOF
-kernel: 4.4.0-87.110-generic
-fully-patched: true
-version: "27.3"
-EOF
-elif [ "$1" = "enable" ]; then
-    echo -n "2017/08/04 18:03:47 Error executing enable?auth-token="
-    echo "deafbeefdeadbeefdeadbeefdeadbeef."
-    echo -n "Machine-token already exists! Please use 'disable' to delete "
-    echo "existing machine-token."
-elif [ "$1" = "disable" ]; then
-    echo -n "Successfully disabled device. Removed machine-token: "
-    echo "deadbeefdeadbeefdeadbeefdeadbeef"
-fi
-exit 0
-"""
-
-LIVEPATCH_DISABLED = """#!/bin/sh
-if [ "$1" = "status" ]; then
-    cat <<EOF
-Machine is not enabled. Please run 'sudo canonical-livepatch enable' with the
-token obtained from https://ubuntu.com/livepatch.
-EOF
-    exit 1
-elif [ "$1" = "enable" ]; then
-    echo -n "Successfully enabled device. Using machine-token: "
-    echo "deadbeefdeadbeefdeadbeefdeadbeef"
-fi
-exit 0
-"""
+from testing import UbuntuAdvantageTest
+from fakes import (
+    SNAP_LIVEPATCH_NOT_INSTALLED,
+    LIVEPATCH_ENABLED,
+    LIVEPATCH_DISABLED)
 
 
-class UbuntuAdvantageTest(TestWithFixtures):
-
-    def setUp(self):
-        super(UbuntuAdvantageTest, self).setUp()
-        tempdir = self.useFixture(TempDir())
-        self.repo_list = Path(tempdir.join('repo.list'))
-        self.bin_dir = Path(tempdir.join('bin'))
-        self.keyrings_dir = Path(tempdir.join('keyrings'))
-        self.trusted_gpg_dir = Path(tempdir.join('trusted.gpg.d'))
-        self.apt_method_https = self.bin_dir / 'apt-method-https'
-        self.ca_certificates = self.bin_dir / 'update-ca-certificates'
-        self.snapd = self.bin_dir / 'snapd'
-        # setup directories and files
-        self.bin_dir.mkdir()
-        self.keyrings_dir.mkdir()
-        self.trusted_gpg_dir.mkdir()
-        (self.keyrings_dir / 'ubuntu-esm-keyring.gpg').write_text('GPG key')
-        self.make_fake_binary('apt-get')
-        self.make_fake_binary('apt-method-https')
-        self.make_fake_binary('update-ca-certificates')
-        self.make_fake_binary('id', command='echo 0')
-        self.make_fake_binary('lsb_release', command='echo precise')
-        self.make_fake_binary('snapd')
-        # in our default setup the snap is installed and enabled
-        self.make_fake_binary('snap', command=SNAP_LIVEPATCH_INSTALLED)
-        self.make_fake_binary('canonical-livepatch', command=LIVEPATCH_ENABLED)
-        self.make_fake_binary('uname', command='echo 4.4.0-89-generic')
-        self.livepatch_token = '0123456789abcdef1234567890abcdef'
-
-    def make_fake_binary(self, binary, command='true'):
-        path = self.bin_dir / binary
-        path.write_text('#!/bin/sh\n{}\n'.format(command))
-        path.chmod(0o755)
-
-    def script(self, *args):
-        """Run the script."""
-        command = ['./ubuntu-advantage']
-        command.extend(args)
-        path = os.pathsep.join([str(self.bin_dir), os.environ['PATH']])
-        env = {
-            'PATH': path,
-            'REPO_LIST': str(self.repo_list),
-            'KEYRINGS_DIR': str(self.keyrings_dir),
-            'APT_KEYS_DIR': str(self.trusted_gpg_dir),
-            'APT_METHOD_HTTPS': str(self.apt_method_https),
-            'CA_CERTIFICATES': str(self.ca_certificates),
-            'SNAPD': str(self.snapd)}
-        process = subprocess.Popen(
-            command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env)
-        process.wait()
-        result = ProcessResult(
-            process.returncode,
-            process.stdout.read().decode('utf-8'),
-            process.stderr.read().decode('utf-8'))
-        process.stdout.close()
-        process.stderr.close()
-        return result
+class UbuntuAdvantageScriptTest(UbuntuAdvantageTest):
 
     def test_enable_disable_needs_root(self):
         """The script must be run as root for enable and disable actions."""
